games/monopoly/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
requests
loguru
Pillow